+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.4.6 (homolog)
 
 * Fixed bug of creating new inbox by chatwoot
+* When conversation reopens is pending when conversation pending is true
 
 # 1.4.5 (2023-07-26 09:32)